+++ js/src/forum/components/TwoFactorItems.tsx
@@ -6,12 +6,27 @@ import type Mithril from "mithril";
 import icon from "flarum/common/helpers/icon"
 import Button from "flarum/common/components/Button";
 import AppSetupModal from "./AppSetupModal";
-
-const trans = (key: string) => {
-  return app.translator.trans(`nearata-twofactor.forum.settings.${key}`);
-};
+import trans from "../helpers/trans";
+import EmailSetupModal from "./EmailSetupModal";
 
 export default class TwoFactorItems extends Component {
+  data = [
+    {
+      key: "app",
+      icon: icon("fas fa-mobile-alt"),
+      title: trans("app_item_label"),
+      desc: trans("app_item_description"),
+      modal: AppSetupModal
+    },
+    {
+      key: "email",
+      icon: icon("fas fa-envelope-open"),
+      title: trans("email_item_label"),
+      desc: trans("email_item_description"),
+      modal: EmailSetupModal
+    }
+  ]
+
   view(_: Mithril.Vnode<this>) {
     return (
       <FieldSet
@@ -19,7 +34,7 @@ export default class TwoFactorItems extends Component {
         label={trans("section_title")}
       >
         <span className="helpText">{trans("section_help")}</span>
-        {this.items().toArray()}
+        <div className="AccessTokensList">{this.items().toArray()}</div>
       </FieldSet>
     );
   }
@@ -27,27 +42,27 @@ export default class TwoFactorItems extends Component {
   items() {
     const items = new ItemList();
 
-    items.add(
-      "app",
-      <div className="AccessTokensList">
+    for (const i of this.data) {
+      items.add(
+        i.key,
         <div className="AccessTokensList-item">
-          <div className="AccessTokensList-item-icon">{icon('fas fa-mobile-alt')}</div>
+          <div className="AccessTokensList-item-icon">{i.icon}</div>
           <div className="AccessTokensList-item-info">
             <div className="AccessTokensList-item-title">
-              <span className="AccessTokensList-item-title-main">{trans('app_item_label')}</span>
+              <span className="AccessTokensList-item-title-main">{i.title}</span>
             </div>
             <div className="AccessTokensList-item-description">
-              <span className="AccessTokensList-item-description-main">{trans('app_item_description')}</span>
+              <span className="AccessTokensList-item-description-main">{i.desc}</span>
             </div>
           </div>
           <div className="AccessTokensList-item-actions">
-            <Button className="Button Button--primary" disabled={false} loading={false} onclick={() => app.modal.show(AppSetupModal)}>
-              {trans('item_manage_label')}
+            <Button className="Button Button--primary" onclick={() => app.modal.show(i.modal)}>
+              {trans("item_manage_label")}
             </Button>
           </div>
         </div>
-      </div>
-    )
+      )
+    }
 
     return items;
   }