+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - admins can no longer disable others 2fa
 - authenticator app: replace FileSaver.js with plain js
 - fix some of the user settings being exposed to others [issue #10](https://github.com/Nearata/flarum-ext-twofactor/issues/15) [pull #15](https://github.com/Nearata/flarum-ext-twofactor/pull/16)
+- add authentication by email [issue #6](https://github.com/Nearata/flarum-ext-twofactor/issues/6)
 
 ## [2.2.1] - 2023-06-18